@@ -2,6 +2,7 @@
 	"name": "next-template",
 	"version": "0.1.0",
 	"private": true,
+	"type": "module",
 	"scripts": {
 		"dev": "next dev",
 		"build": "next build",
@@ -10,17 +11,20 @@
 	},
 	"dependencies": {
 		"@whop-sdk/core": "^0.0.7",
-		"next": "13.2.4",
-		"next-auth": "^4.18.0",
-		"react": "18.2.0",
-		"react-dom": "18.2.0"
+		"next": "^14.2.4",
+		"next-auth": "^4.24.7",
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1"
 	},
 	"devDependencies": {
-		"@types/node": "18.11.11",
-		"@types/react": "18.0.26",
-		"@types/react-dom": "18.0.9",
-		"typescript": "4.9.3",
-		"eslint": "8.29.0",
-		"eslint-config-next": "13.0.6"
+		"eslint": "^8.57.0",
+		"eslint-config-next": "13.0.6",
+		"@types/node": "^20.14.10",
+		"@types/react": "^18.3.3",
+		"@types/react-dom": "^18.3.0",
+		"@types/eslint": "^8.56.10",
+		"@typescript-eslint/eslint-plugin": "^7.1.1",
+		"@typescript-eslint/parser": "^7.1.1",
+		"typescript": "^5.5.3"
 	}
 }
